requirements.txt
flet==0.23.*
python-dotenv~=1.0.1
psycopg2~=2.9.9
